Various changes to routes, fixes gh-1
There was an issue with a SQL statement regarding the update of
routes which gets fixed with this commit. Also some changes have
been made to the way routes get displayed, most notebly the font
has been slightly adjusted, and in the TableModel the conversion
of Strings to Integer has been improved.