Preselection of main hub when creating routes, closes gh-2
When a new route gets created, the main hub of an airline is now
the suggested selection for the origin. To provide this feature
several changes have been made to the way enterprises are handled,
and to the flow of dialogs. Most notebly an enterprise does not
have an home country anymore, but a main hub from where it operates.
Accordingly the dialog for creating an enterprise has changed, and
of course the route creation dialog itself.