Table updates changes to model automatically
The table model has been adjusted to update its rows automatically
when a value gets modified. E.g. if several routes to an airport
exist and one value of it gets changed, this change is immediatly
displayed in all rows where this airport is listed. Before this the
update took place only when the user selected the accordings rows
manually.