Option added to delete routes
The user can now delete rows by pressing the delete key. This removes
the currently selected row, and deletes the route associated with it
from the database.

Furthermore some changes have been made to the way airports are
deleted. Contratry to previous statements airports cannot be
deleted if one or more routes still exist which use this airport
as the start or destination.